Feedback
Hey Pal, 

Overall this is great work. I really liked how you were thinking about ways to clean up the code and make it more efficient.  I have some ideas which I think will push your idea about efficient code even further.

Functions: The sorting and mapping functions are great. They do a good job of simplifying the code in the fetchProjects function. However, these functions are doing more than you are suggesting in the function names. Basically, what is happening is you are taking repeating code and placing it in a function which is good because it reduces code but we want to make sure we aren’t moving code around just to have it somewhere else.  If we take the sorting function as an example, we actually already have a sorting function below (the hondleSort function). This function could be expanded to help sort the project creation date by ascending which is what your sort function does. So rather than repeat a sort function we can expand one that already exists. As for the mapping function this only solves the fact that the same block of code is repeated twice but it is kind of combining 3 bits of functionality. 1. Mapping over an array of guest codes 2. Retrieving a project based on a guest code and 3. Adding a new project to array. So in my mind what I might do is keep the mapping function so that it maps over an array of guestCodes. So you might want to call it something like mapGuestCodes instead, or just leave the mapping out of a function and put it back in the fetchProjects code. Next you could create two new more generalized functions. One which retrieves a project based on a guestCode and another which adds an object to an array. Then place the setProjects part back in the fetchProjects function.

Really what I’m trying to say is that if you decide to create a new function it should be for something more generalized if possible. Reducing repeated code is good but ideally you also set yourself up to be able to use the function for future repeated code as well and the best way to do that is make it more generalized.

Comments: You've got good coverage of your comments, meaning you've got comments in the places I'd expect them but your commenting/documentation is a bit generic at the moment. Not a bad thing! It's a good starting point. What I'd like you try is to explain what is actually happening in these functions. For example give a better explanation of what data is returned by some of the API calls, what type of actions are performed in the function and most importantly what the expected outcome of the function is. This is a bit overkill for normal documentation but this is meant to be an exercise to familiarize yourself with this codebase. Using console.log() will be your friend. If you'd like to see how I would approach this, I can walk through one as an example with you on Monday. 

Overall good start Pal. But I’m going to push you to take it even further. Generalize the functions as much as you can. Also increase detail in the documentation so that if you or I come back to this code in 2 months and look at it we can just read your documentation to understand what we were trying to achieve.